worker/jobs/index: Extract sync jobs into dedicated module
